refactor: audit backlog #2402 — IAuditLog.IsPersisted + split import job service file
Two small, contained cohesion fixes from the 2026-07 platform cross-cut audit
(honua-server#2402), triaged as low-risk and safe for a release cut:

- PA-041: DefaultComplianceDependencyGate detected the null audit-log
  implementation by string-matching the concrete type name
  ("NullAuditLog"), piercing the IAuditLog abstraction. Added an
  IsPersisted capability property to IAuditLog (default true via a
  default interface member, so existing implementations and test doubles
  are unaffected) and overridden to false only in NullAuditLog. The gate
  now asks the abstraction instead of reflecting on the concrete type.

- PA-038: InMemoryImportJobService.cs bundled two unrelated
  IImportJobService implementations (InMemoryImportJobService and
  UniversalImportJobService) in one 1165-line file. Split
  UniversalImportJobService (and its private nested types) into its own
  UniversalImportJobService.cs file with a pruned using list; no logic
  changes. The two ImportJobState nested types are not true duplicates
  (different shapes/lifecycles per implementation) so consolidating them
  is left for a follow-up rather than risked here.

The remaining 33 findings in the audit are god classes / large
service-locator cleanups spanning many files and DI registrations
(some 2,000-14,000 lines); those are deferred as out of scope for a
low-risk pass near a release cut and tracked on the issue.
